\name{TERAplusB-package}
\alias{TERAplusB-package}
\alias{TERAplusB}
\docType{package}
\title{
A+B Traditional Escalation Rule
}
\description{
TERAplusB is designed for the comparison of various A+B TER.
}

\author{ Eun-Kyung Lee}

\references{
...
}

\keyword{ utilities }

\examples{
TER.deesc.A.B(c("D1","D2"),c(0.2,0.7),3,3,1,2,3)
}